إصلاح: حل جميع أخطاء Pyright ومشاكل التايبينغ — صفر مشاكل
- إضافة typing import وتوسيع توقيع send_message في communications/services.py لدعم النمط المباشر (recipient_phone/channel_type/context/user_id)
- استبدال with transaction.atomic() بـ typing.cast لحل خطأ Context Manager في Pyright في communications/services.py وprovisioning.py
- إصلاح استدعاء process_campaign_task.delay باستخدام getattr لتجنب خطأ Pyright
- إصلاح actual_subject | None عند تمريره لـ _branded_email_html
- نقل import typing إلى أعلى provisioning.py
- إصلاح CommunicationAttachment.__str__ و CommunicationStatistics.__str__ بإضافة str() وإصلاح وصول حقل التاريخ
- إصلاح StudentProfile.__str__ في students/domain/models.py بإرجاع str(self.arabic_name)
- إضافة logging.getLogger('nebras.students') في students/application/services.py لحل خطأ 'Could not find name logger'
- إصلاح سطر std_name في create_student_manually باستخدام profile_data.get('arabic_name') بدلاً من المتغير غير المعرّف
- استبدال User.DoesNotExist و Voucher.DoesNotExist و JournalEntry.DoesNotExist بـ ObjectDoesNotExist في student_finance/services.py
- إصلاح finance/services.py بإضافة @transaction.atomic # type: ignore[arg-type] على post_journal_entry و reverse_journal_entry

Django check: 0 مشاكل | TypeScript: 0 أخطاء

+++ backend/apps/common/utils/tafqeet.py
@@ -32,8 +32,8 @@ def tafqeet_arabic(num, currency: str = "جنيه سوداني") -> str:
     hundreds = ["", "مائة", "مائتان", "ثلاثمائة", "أربعمائة", "خمسمائة", "ستمائة", "سبعمائة", "ثمانمائة", "تسعمائة"]
     scales = ["", "ألف", "مليون", "مليار", "تريليون"]
 
-    int_part = int(math.floor(val))
-    frac_part = int(round((val - int_part) * 100))
+    int_part = math.floor(val)
+    frac_part = round((val - int_part) * 100)
 
     def convert_3_digits(n: int) -> str:
         if n == 0:
